# docs/swagger.yaml
definitions:
...
  domain.Employee:
    properties:
      card_number_id:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      warehouse_id:
        type: integer
    type: object
  domain.EmployeeRequest:
    properties:
      card_number_id:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      warehouse_id:
        type: integer
    required:
    - card_number_id
    - first_name
    - last_name
    - warehouse_id
    type: object
  domain.EmployeeWithInboundOrders:
    properties:
      card_number_id:
        type: string
      first_name:
        type: string
      id:
        type: integer
      inbound_orders_count:
        type: integer
      last_name:
        type: string
      warehouse_id:
        type: integer
    type: object
...
  domain.Locality:
    properties:
      country_name:
        type: string
      id:
        type: string
      locality_name:
        type: string
      province_name:
        type: string
    required:
    - country_name
    - id
    - locality_name
    - province_name
    type: object
...
  domain.QuantitySellerByLocality:
    properties:
      locality_id:
        type: string
      locality_name:
        type: string
      sellers_count:
        type: integer
    type: object
...
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact: {}
  description: This API was created as part of Mercado Libre's Go backend bootcamp
    for training purposes.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es-ar/terminos-y-condiciones
  title: MeLi Bootcamp API
  version: "1.0"
paths:
...
  /api/v1/employees:
    get:
      description: get employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Employee'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: List employees
      tags:
      - Employees
    post:
      consumes:
      - application/json
      description: create employee
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.EmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Employee'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create employee
      tags:
      - Employees
  /api/v1/employees/{id}:
    delete:
      description: delete employee by id
      parameters:
      - description: Employee Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete employee
      tags:
      - Employees
    get:
      description: get employee by id
      parameters:
      - description: Employee Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Employee'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get employee
      tags:
      - Employees
    patch:
      consumes:
      - application/json
      description: update employee
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.EmployeeRequest'
      - description: Employee Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Employee'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Update employee
      tags:
      - Employees
  /api/v1/employees/reportInboundOrders:
    get:
      description: get employee with inbound orders count
      parameters:
      - description: Employee Id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.EmployeeWithInboundOrders'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Employee with inbound orders count
      tags:
      - Employees
...
  /api/v1/localities:
    post:
      consumes:
      - application/json
      description: Create locality
      parameters:
      - description: Locality parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Locality'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Locality'
              type: object
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create locality
      tags:
      - Localities
...
  /api/v1/localities/reportSellers:
    get:
      description: Returns a list of all reports
      parameters:
      - description: locality Id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.QuantitySellerByLocality'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: ReportSellers
      tags:
      - Localities
...
securityDefinitions:
  ApiKeyAuth:
    description: start with Bearer
    in: header
    name: x-tiger-token
    type: apiKey
swagger: "2.0"